Possible fix for infinite load issues
